Putting get border cell before allgather onto the gpu

+++ clamr.cpp
@@ -473,7 +473,7 @@ extern "C" void do_calc(void)
 
       //  Resize the mesh, inserting cells where refinement is necessary.
       cpu_timer_start(&tstart_cpu);
-      if (ncells_global_old != ncells_global) state->gpu_rezone_all(command_queue, mesh, old_ncells, new_ncells, old_ncells, localStencil);
+      if (ncells_global_old != (int)ncells_global) state->gpu_rezone_all(command_queue, mesh, old_ncells, new_ncells, old_ncells, localStencil);
       cpu_time_rezone += cpu_timer_stop(tstart_cpu);
 
       // XXX XXX XXX
